// api/playauto_api_type.ts
export interface IPAJobCallbackProcessingResponse {
    job_id: string
    current: number
    total: number
    state_msg: string
    job_cate: string
    sol_code: string
}
export interface IPAJobCallbackDoneResponse<T> {
    job_id: string
    title: string
    results: IPAJobCallbackResults<T>
}

export type IPAJobCallbackResponse<T> = IPAJobCallbackProcessingResponse | IPAJobCallbackDoneResponse<T>;

export interface IPAJobCallbackResults<T> {
    "config.json": IPAJobCallbackConfigJson
    "result.json": T[] | IPAJobCallbackFailedResultJson
}

export interface IPAJobCallbackFailedResultJson {
    Result: string
    Message: string
}

export interface IPAJobCallbackConfigJson {
    sol_code: string
}

export interface IPAJobCallbackRegistProdResultJson {
    state: number
    site_code: string
    site_id: string
    code: string
    sku_code: any
    single_yn: any
    groupkey: any
    slave_reg_code: string
    slave_reg_code_sub: string
    reg_type: string
    reg_sell_term: number
    reg_fee: number
    reg_premium: number
    slave_wdate: string
    slave_edate: string
    msg: string
    err_code: any
    setdata: string
    setName: string
    toState: string
    gmp_sale_no: any
    isSavedToDB: boolean
}